fix(dev-desktop): address second review pass on multi-run support
- build-main-bundle.cjs: loop over bundle output dirs instead of
  repeating rmSync/mkdirSync per entry
- dev-stack.cjs: handle spawn errors instead of leaving them uncaught
- move dev-renderer-origin resolution out of electron-main/ into
  ipc-contracts/ so the dev-seed preload no longer imports main-only
  modules
- host-auth.ts: run deleteCredentials() before any slot-aware logger/
  path resolution, so a malformed DEV_DESKTOP_SLOT can no longer skip
  credential clearing
- label.ts: give ServiceLabel an explicit devSlot field instead of
  windowsTaskName re-parsing it out of label.id
- consolidate the withDevDesktopSlot test helper (duplicated across 4
  test files in two workspaces) into clients/shared/test-fixtures
- paths.test.ts: assert cliLogPath alongside the other dev-desktop-slot
  install surfaces; fix the doc comment that omitted it
- dev-desktop.test.mjs: clean up temp host-home dirs in finally blocks
  so a failing assertion doesn't leak them
