+++ clients/desktop/src/electron-main/dev-desktop-runtime.ts
@@ -1,54 +1,10 @@
 import type { Environment } from "../config";
 import { devDesktopSlotForEnvironment } from "./host/dev-desktop-slot";
 
-export const TRAYCER_DESKTOP_DEV_URL_ENV = "TRAYCER_DESKTOP_DEV_URL";
-export const DEFAULT_DEV_RENDERER_URL = "http://localhost:5173";
-
-const ALLOWED_DEV_RENDERER_HOSTS: ReadonlySet<string> = new Set([
-  "localhost",
-  "127.0.0.1",
-]);
-
 export interface DesktopRuntimeIdentity {
   readonly appName: string;
   readonly userDataDirName: string | null;
   readonly slot: string | null;
-}
-
-export function devRendererUrlFromEnv(env: NodeJS.ProcessEnv): string {
-  const raw = env[TRAYCER_DESKTOP_DEV_URL_ENV] ?? DEFAULT_DEV_RENDERER_URL;
-  return normalizeDevRendererUrl(raw);
-}
-
-export function devRendererOriginFromEnv(env: NodeJS.ProcessEnv): string {
-  return new URL(devRendererUrlFromEnv(env)).origin;
-}
-
-export function normalizeDevRendererUrl(value: string): string {
-  let url: URL;
-  try {
-    url = new URL(value);
-  } catch {
-    throw new Error(`${TRAYCER_DESKTOP_DEV_URL_ENV} must be a valid URL`);
-  }
-  if (url.protocol !== "http:") {
-    throw new Error(`${TRAYCER_DESKTOP_DEV_URL_ENV} must use http`);
-  }
-  if (!ALLOWED_DEV_RENDERER_HOSTS.has(url.hostname)) {
-    throw new Error(`${TRAYCER_DESKTOP_DEV_URL_ENV} must use a loopback host`);
-  }
-  if (url.port.length === 0) {
-    throw new Error(`${TRAYCER_DESKTOP_DEV_URL_ENV} must include a port`);
-  }
-  if (url.username.length > 0 || url.password.length > 0) {
-    throw new Error(
-      `${TRAYCER_DESKTOP_DEV_URL_ENV} must not include credentials`,
-    );
-  }
-  if (url.pathname !== "/" || url.search.length > 0 || url.hash.length > 0) {
-    throw new Error(`${TRAYCER_DESKTOP_DEV_URL_ENV} must be an origin URL`);
-  }
-  return url.origin;
 }
 
 export function resolveDesktopRuntimeIdentity(
