+++ clients/traycer-cli/src/host/__tests__/auto-bootstrap.test.ts
@@ -149,6 +149,7 @@ beforeEach(() => {
       id: environment === "dev" ? "ai.traycer.host.dev" : "ai.traycer.host",
       displayName: "Traycer Host",
       environment,
+      devSlot: null,
     }),
   );
   resolveServiceCliInvocationMock.mockReturnValue({
