+++ clients/traycer-cli/src/service/label.ts
@@ -24,12 +24,17 @@ export interface ServiceLabel {
   readonly displayName: string;
   // Exact runtime environment this label operates on.
   readonly environment: Environment;
+  // The dev-desktop run slot baked into `id`/`displayName`, or `null` when
+  // this label isn't slot-specific. First-class so consumers (e.g.
+  // `windowsTaskName`) never have to re-derive it by re-parsing `id`.
+  readonly devSlot: string | null;
 }
 
 const PRODUCTION_LABEL: ServiceLabel = {
   id: "ai.traycer.host",
   displayName: "Traycer Host",
   environment: "production",
+  devSlot: null,
 };
 
 // Each non-production environment gets its OWN service slot
@@ -52,12 +57,14 @@ export function serviceLabelFor(environment: Environment): ServiceLabel {
       id: `ai.traycer.host.dev.${devSlot}`,
       displayName: `Traycer Host (Dev ${devSlot})`,
       environment,
+      devSlot,
     };
   }
   return {
     id: `ai.traycer.host.${environment}`,
     displayName: `Traycer Host (${capitalizeEnvironment(environment)})`,
     environment,
+    devSlot: null,
   };
 }
 
@@ -82,9 +89,8 @@ export function serviceManifestPath(label: ServiceLabel): string {
 // `Host-Staging`).
 export function windowsTaskName(label: ServiceLabel): string {
   if (label.environment === "production") return "\\Traycer\\Host";
-  const devPrefix = "ai.traycer.host.dev.";
-  if (label.id.startsWith(devPrefix)) {
-    return `\\Traycer\\Host-Dev-${capitalizeEnvironment(label.id.slice(devPrefix.length))}`;
+  if (label.devSlot !== null) {
+    return `\\Traycer\\Host-Dev-${capitalizeEnvironment(label.devSlot)}`;
   }
   return `\\Traycer\\Host-${capitalizeEnvironment(label.environment)}`;
 }
