+++ clients/traycer-cli/src/upgrade/__tests__/finalize-helper.test.ts
@@ -100,6 +100,7 @@ describe("scheduleFinalizationHelper", () => {
         id: "ai.traycer.host.prod",
         displayName: "Traycer Host",
         environment: "production",
+        devSlot: null,
       },
       parentPid: 4242,
       parentExitTimeoutSeconds: 60,
@@ -162,6 +163,7 @@ describe("scheduleFinalizationHelper", () => {
         id: "ai.traycer.host.prod",
         displayName: "Traycer Host",
         environment: "production",
+        devSlot: null,
       },
       parentPid: 4242,
       parentExitTimeoutSeconds: 60,
@@ -195,6 +197,7 @@ describe("scheduleFinalizationHelper", () => {
         id: "ai.traycer.host.prod",
         displayName: "Traycer Host",
         environment: "production",
+        devSlot: null,
       },
       parentPid: 4242,
       parentExitTimeoutSeconds: 60,
@@ -222,6 +225,7 @@ describe("scheduleFinalizationHelper", () => {
         id: "ai.traycer.host.prod",
         displayName: "Traycer Host",
         environment: "production",
+        devSlot: null,
       },
       parentPid: 4242,
       parentExitTimeoutSeconds: 60,
@@ -262,6 +266,7 @@ describe("scheduleFinalizationHelper", () => {
         id: "ai.traycer.host.prod",
         displayName: "Traycer Host",
         environment: "production",
+        devSlot: null,
       },
       parentPid: 4242,
       parentExitTimeoutSeconds: 60,
